Починен WeatherApi (Использование Call вместо Response) + ext функции для подписки на flow + доработка логики получения погоды + правка Network модели

+++ app/src/main/java/ru/upsoft/gpspointer/data/network/weather/WeatherNetworkModel.kt
@@ -45,7 +45,7 @@ data class Main(
 
 data class Wind(
 
-    @SerializedName("speed") val speed: Int,
+    @SerializedName("speed") val speed: Float,
     @SerializedName("deg") val deg: Int
 )
